fix: memory leak in all useEffect statements with isMounted flag, also swap user logic for axios call in Profile.js
